reduce the size of juicefs.lite
