+++ pkg/meta/tkv_tikv.go
@@ -1,3 +1,5 @@
+// +build !notikv
+
 /*
  * JuiceFS, Copyright (C) 2021 Juicedata, Inc.
  *
@@ -17,7 +19,6 @@ package meta
 
 import (
 	"context"
-	"fmt"
 	"strings"
 
 	tikverr "github.com/tikv/client-go/v2/error"
@@ -26,18 +27,23 @@ import (
 
 func init() {
 	Register("tikv", newKVMeta)
+	drivers["tikv"] = newTikvClient
+
 }
 
-func newTkvClient(driver, addr string) (tkvClient, error) {
-	if driver == "memkv" {
-		return newMockClient()
-	}
-	if driver != "tikv" {
-		return nil, fmt.Errorf("invalid driver %s != expected %s", driver, "tikv")
+func newTikvClient(addr string) (tkvClient, error) {
+	p := strings.Index(addr, "/")
+	var prefix string
+	if p > 0 {
+		prefix = addr[p+1:]
+		addr = addr[:p]
 	}
 	pds := strings.Split(addr, ",")
 	client, err := tikv.NewTxnClient(pds)
-	return &tikvClient{client}, err
+	if err != nil {
+		return nil, err
+	}
+	return withPrefix(&tikvClient{client}, append([]byte(prefix), 0xFD)), nil
 }
 
 type tikvTxn struct {
