+++ pkg/object/tikv.go
@@ -1,3 +1,5 @@
+// +build !notikv
+
 /*
  * JuiceFS, Copyright (C) 2021 Juicedata, Inc.
  *
